% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biochemistry-tables}
\alias{biochemistry-tables}
\alias{table_Vcmax_}
\alias{table_Jmax_}
\title{Biochemical model lookup tables for Vcmax25/Jmax temperature response}
\format{
A data frame of temperature-response coefficients.

An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
table_Vcmax_

table_Jmax_
}
\description{
Lookup tables of the temperature-response parameters used by the
biochemical (photosynthesis) model to scale maximum carboxylation
(\code{Vcmax25}) and electron transport (\code{Jmax}) capacity with leaf
temperature.
}
\keyword{datasets}
