% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brdf-angle-datasets}
\alias{brdf-angle-datasets}
\alias{brdf_angles}
\alias{brdf_angles2}
\alias{brdf_angles_no_oversampling}
\title{BRDF viewing/illumination angle configurations}
\format{
A data frame with columns for solar zenith, viewing zenith and relative azimuth angles (degrees).

An object of class \code{data.frame} with 333 rows and 2 columns.

An object of class \code{data.frame} with 55 rows and 2 columns.
}
\usage{
brdf_angles

brdf_angles2

brdf_angles_no_oversampling
}
\description{
Predefined sets of solar/viewing zenith and relative azimuth angles used
to compute the bidirectional reflectance distribution function (BRDF) of
the canopy. \code{brdf_angles_no_oversampling} omits the finer angular
oversampling used by \code{brdf_angles}/\code{brdf_angles2}.
}
\keyword{datasets}
