% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calczenithangle.R
\name{calczenithangle}
\alias{calczenithangle}
\title{Calculates pi/2-the angle of the sun with the slope of the surface.}
\usage{
calczenithangle(Doy, t, Omega_g, Fi_gm, Long, Lat)
}
\arguments{
\item{Doy}{day of the year}

\item{t}{time of the day (hours, GMT)}

\item{Omega_g}{slope azimuth angle (deg)}

\item{Fi_gm}{slope of the surface (deg)}

\item{Long}{Longitude (decimal)}

\item{Lat}{Latitude (decimal)}
}
\value{
A list \code{zenith.angles} intended to contain:
\describe{
\item{Omega_s}{Hour angle of the sun (rad).}
\item{Fi_s}{"Classic" solar zenith angle, perpendicular to the horizontal plane (rad).}
\item{Fi_g}{Projected slope of the surface, in the plane through the solar beam and the vertical (rad).}
\item{Fi_gs}{Angle of the sun with the vector perpendicular to the (possibly sloped) surface, i.e. \code{pi/2} minus the surface solar elevation angle (rad).}
}
}
\description{
Calculates pi/2-the angle of the sun with the slope of the surface.
}
\details{
Last updates: Jan 2003; Oct 2008 by Joris Timmermans (corrected equation of time); Oct 2012 (CvdT) comment: input time is GMT, not local time.
}
\examples{
\dontrun{
zenith.angles <- calczenithangle(Doy = 180, t = 12, Omega_g = 210,
                                  Fi_gm = 30, Long = 13.75, Lat = 45.5)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol  (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
