% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluspect_functions.R
\name{define.constants}
\alias{define.constants}
\title{Define physical constants used by SCOPEinR}
\usage{
define.constants()
}
\value{
A list with named numeric elements: \code{A} (Avogadro's number, mol^-1), \code{h} (Planck's constant, J s), \code{c} (speed of light, m s^-1), \code{cp} (specific heat of dry air, J kg^-1 K^-1), \code{R} (molar gas constant, J mol^-1 K^-1), \code{rhoa} (specific mass of air, kg m^-3), \code{g} (gravitational acceleration, m s^-2), \code{kappa} (Von Karman constant, dimensionless), \code{MH2O} (molecular mass of water, g mol^-1), \code{Mair} (molecular mass of dry air, g mol^-1), \code{MCO2} (molecular mass of CO2, g mol^-1), \code{sigmaSB} (Stefan-Boltzmann constant, W m^-2 K^-4), \code{deg2rad} (degrees-to-radians conversion factor), \code{C2K} (Celsius-to-Kelvin offset, K).
}
\description{
\code{define.constants} returns a list of the physical constants used
throughout the SCOPEinR model (radiative transfer, energy balance and
biochemical sub-models), analogous to the \code{SCOPEinR::constants} table.
}
\examples{
const <- define.constants()
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Wout Verhoef (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
