% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_temp_response_biochem.R
\name{define_temp_response_biochem}
\alias{define_temp_response_biochem}
\title{get the variables needed in temp_response_biochem}
\usage{
define_temp_response_biochem(getTDP = T)
}
\arguments{
\item{getTDP}{is TRUE get a list with TDP}
}
\value{
get a list with all parameters for TDP
}
\description{
get the variables needed in temp_response_biochem
}
\examples{
TDP <- define_temp_response_biochem(getTDP = TRUE)
}
