% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{get.BallBerry}
\alias{get.BallBerry}
\title{get.BallBerry
\code{get.BallBerry}   get BerryBall value from Berry Model}
\usage{
get.BallBerry(Cs, RH, A, BallBerrySlope, BallBerry0, minCi, Ci_input)
}
\arguments{
\item{Cs}{is CO2 at leaf surface}

\item{RH}{is relative humidity}

\item{A}{is Net assimilation in 'same units of CO2 as Cs'/m2/s}

\item{BallBerrySlope}{parameter for  BallBerry model}

\item{BallBerry0}{parameter for  BallBerry model}

\item{minCi}{minimum Ci as a fraction of Cs (in case RH is very low?)}

\item{Ci_input}{will use only for Ci == Ci_input}
}
\value{
A list with gs (stomatal conductance) and Ci (intercellular CO2 concentration).
}
\description{
get.BallBerry
\code{get.BallBerry}   get BerryBall value from Berry Model
}
\examples{
get.BallBerry(Cs = 400, RH = 0.7, A = 15, BallBerrySlope = 9, BallBerry0 = 0.01, minCi = 0.3)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
