% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biochemical.R
\name{get.biochemical}
\alias{get.biochemical}
\title{get.biochemical}
\usage{
get.biochemical(data.leafbio, data.meteo, data.opts, fV, get.plots)
}
\arguments{
\item{data.leafbio}{list. Leaf biochemical parameters: \code{Type} ('C3' or 'C4'), \code{stressfactor}, \code{Vcmax25}, \code{BallBerry0}, \code{BallBerrySlope}, \code{Rdparam}, \code{Kn0}/\code{Knalpha}/\code{Knbeta} (NPQ model parameters), optional \code{g_m} (mesophyll conductance, mol m^-2 s^-1 bar^-1), and (when temperature correction is enabled) the \code{TDP} sub-list of temperature-dependence parameters.}

\item{data.meteo}{list. Meteorological/leaf micro-environment characteristics: \code{Q} (absorbed PAR, umol photons m^-2 s^-1), \code{Cs} (CO2 concentration in the leaf boundary layer, ppm), \code{Temp} (leaf temperature, deg C or K), \code{eb} (vapour pressure in the leaf boundary layer, hPa), \code{Oa} (O2 concentration, mmol/mol), \code{p} (air pressure, hPa).}

\item{data.opts}{data.frame. Simulation options; row 7 (\code{tempcor}) is a 0/1 flag for whether temperature correction of Vcmax and the rate constants is applied.}

\item{fV}{numeric. Scaling factor applied to \code{Vcmax25} (e.g. the canopy nitrogen/Vcmax profile factor estimated in the energy balance).}

\item{get.plots}{logical. If \code{TRUE}, returns diagnostic plots for stomatal conductance, assimilation, and Jmax/Vcmax rates.}
}
\value{
A list \code{biochem_out} with (among others): \code{A} (net assimilation rate, umol m^-2 s^-1), \code{Ci}/\code{Cc} (internal/chloroplast CO2 concentration, ppm), \code{rcw} (stomatal resistance, s m^-1), \code{gs} (stomatal conductance), \code{Vcmax}, \code{Rd}, \code{Ja} (actual electron transport rate), \code{ps}/\code{ps_rel} (photochemical yield and degree of light saturation), \code{eta} (fluorescence yield relative to dark-adapted), \code{qE}/\code{qQ} (non-photochemical/photochemical quenching), \code{fs}/\code{fo}/\code{fm} (steady-state/dark/light-saturated fluorescence yields), \code{Kn}/\code{NPQ}/\code{Kf}/\code{Kp}/\code{Kd} (rate constants), and \code{SIF} (solar-induced fluorescence, \code{fs * Q}).
}
\description{
\code{get.biochemical} Calculates net assimilation rate A, fluorescence F using biochemical model

This function calculates:
\itemize{
\item stomatal resistance of a leaf or needle (s m-1)
\item photosynthesis of a leaf or needle (umol m-2 s-1)
\item fluorescence of a leaf or needle (fraction of fluor. in the dark)
}
}
\examples{
\dontrun{
biochem_out <- get.biochemical(data.leafbio, data.meteo, data.opts, fV = 1, get.plots = FALSE)
}
}
\references{
Farquhar et al. 1980, Collatz et al (1991, 1992), and:
Dutta, D., Schimel, D. S., Sun, Y., Tol, C. V. D., & Frankenberg, C. (2019).
Optimal inverse estimation of ecosystem parameters from observations of carbon and energy fluxes.
Biogeosciences, 16(1), 77-103.

Van der Tol, C., Berry, J. A., Campbell, P. K. E., & Rascher, U. (2014).
Models of fluorescence and photosynthesis for interpreting measurements
of solar induced chlorophyll fluorescence. Journal of Geophysical Research: Biogeosciences, 119(12), 2312-2327.

Bonan, G. B., Lawrence, P. J., Oleson, K. W., Levis, S., Jung, M., Reichstein, M., ... & Swenson, S. C. (2011).
Improving canopy processes in the Community Land Model version 4 (CLM4)
using global flux fields empirically inferred from FLUXNET data.
Journal of Geophysical Research: Biogeosciences, 116(G2).
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Joe Berry and Christiaan van der Tol, Ari Kornfeld (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)

last updates
Date: 	21 Sep 2012.

Update:   20 Feb 2013.
Update:      Aug 2013: correction of L171: Ci <- Ci*1e6/ p * 1E3

Update:   2016-10 - (JAK) major rewrite to accomodate an iterative solution to the Ball-Berry equation
- also allows for g_m to be specified for C3 plants, but only if Ci_input is provided.

Update:   25 Feb 2021: Temperature reponse functions by Dutta et al. implemented
}
