% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{get.Ci.next}
\alias{get.Ci.next}
\title{get.Ci.next
\code{get.Ci.next} Function to calculate the difference between "guessed" Ci (Ci_in) and Ci computed using BB after computing A
Test-function for iteration (note that it assigns A in the function's context.)
As with the next section, this code can be read as if the function body executed at this point. (if iteration was used).
In other words, A is assigned at this point in the file (when iterating).}
\usage{
get.Ci.next(Ci_in, Cs, RH, minCi, BallBerrySlope, BallBerry0, A_fun, ppm2bar)
}
\arguments{
\item{Ci_in}{numeric. The "guessed" intercellular CO2 concentration to test.}

\item{Cs}{numeric. CO2 concentration at the leaf surface.}

\item{RH}{numeric. Relative humidity (0-1).}

\item{minCi}{numeric. Minimum Ci as a fraction of Cs.}

\item{BallBerrySlope}{numeric. Slope parameter of the Ball-Berry stomatal conductance model.}

\item{BallBerry0}{numeric. Intercept (minimum conductance) parameter of the Ball-Berry model.}

\item{A_fun}{function. Assimilation function, called as \code{A_fun(Cs)} to compute net assimilation.}

\item{ppm2bar}{numeric. Conversion factor from ppm to bar for CO2 partial pressure.}
}
\value{
A list with err (difference between guessed and Ball-Berry-computed Ci) and Ci_out (the Ball-Berry result).
}
\description{
get.Ci.next
\code{get.Ci.next} Function to calculate the difference between "guessed" Ci (Ci_in) and Ci computed using BB after computing A
Test-function for iteration (note that it assigns A in the function's context.)
As with the next section, this code can be read as if the function body executed at this point. (if iteration was used).
In other words, A is assigned at this point in the file (when iterating).
}
\examples{
\dontrun{
get.Ci.next(Ci_in = 280, Cs = 400, RH = 0.7, minCi = 0.3, BallBerrySlope = 9,
            BallBerry0 = 0.01, A_fun = function(Cs) list(A = 15), ppm2bar = 1e-6)
}
}
