% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2phot.R
\name{get.e2phot}
\alias{get.e2phot}
\title{get.e2phot calculates the number of moles of photons}
\usage{
get.e2phot(lambda, E, constants)
}
\arguments{
\item{lambda}{numeric vector. Wavelength(s) of the photons (m).}

\item{E}{numeric vector. Radiant energy at each wavelength (J), same length as \code{lambda}.}

\item{constants}{data.frame. Physical constants table (as used throughout SCOPEinR), must contain rows for \code{h} (Planck's constant, J s), \code{c} (speed of light, m s^-1) and \code{A} (Avogadro's number, mol^-1) in columns \code{constant}/\code{value}.}
}
\value{
numeric vector. Number of moles of photons (mol) corresponding to \code{E} Joules of energy at each wavelength \code{lambda}.
}
\description{
\code{get.e2phot} calculates the number of moles of photons
corresponding to E Joules of energy of wavelength lambda (m)
}
\examples{
\dontrun{
molphotons <- get.e2phot(lambda, E, constants)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Wout Verhoef, Christiaan van der Tol, Joris Timmermans (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
