% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebal.R
\name{get.ebal}
\alias{get.ebal}
\title{get.ebal
\code{get.ebal} Calculates the energy balance of a vegetated surface}
\usage{
get.ebal(
  data.rad,
  data.gap,
  data.meteo,
  data.soil,
  data.canopy,
  data.leafbio,
  data.leafopt,
  data.spectral,
  data.opts,
  integrate.layer,
  k.maxit,
  get.plots
)
}
\arguments{
\item{data.rad}{incident radiation}

\item{data.gap}{probabilities of direct light penetration and viewing}

\item{data.meteo}{meteo characterisitics}

\item{data.soil}{soil properties}

\item{data.canopy}{canopy properties}

\item{data.leafbio}{leaf biochemical parameters}

\item{data.leafopt}{leaf biochemical parameters}

\item{data.spectral}{spectral information for the model}

\item{data.opts}{options for running the model}

\item{integrate.layer}{how estimate the sum of layer (angles, layers or angles_layers)}

\item{k.maxit}{maximum number for iteration on the balance model}

\item{get.plots}{is true plot the intermediate plots}
}
\value{
a list with:
\itemize{
\item iter        numerical parameters used in the iteration for energy balance closure.
\item fluxes      energy balance, turbulent, and CO2 fluxes.
\item rad         radiation spectra.
\item thermal     temperatures, aerodynamic resistances and friction velocity.
\item bcu, bch    leaf biochemical outputs for sunlit and shaded leaves,
}
}
\description{
date   26 Nov 2007 (CvdT)
updates:
\itemize{
\item 29 Jan 2008 (JT & CvdT)     converted into a function
\item 11 Feb 2008 (JT & CvdT)     improved soil heat flux and temperature calculation
\item 14 Feb 2008 (JT)            changed h in to hc (as h=Avogadro`s constant)
\item 31 Jul 2008 (CvdT)          Included Pntot in output
\item 19 Sep 2008 (CvdT)          Converted F0 and F1 from units per aPAR into units per iPAR
\item 07 Nov 2008 (CvdT)          Changed layout
\item 18 Sep 2012 (CvdT)          Changed Oc, Cc, ec
\item Feb 2012 (WV)            introduced structures for variables
\item Sep 2013 (JV, CvT)       introduced additional biochemical model
\item 10 Dec 2019 (CvdT)          made a light version (layer averaged fluxes)
}
}
\examples{
\dontrun{
out <- get.ebal(data.rad, data.gap, data.meteo, data.soil, data.canopy,
                 data.leafbio, data.leafopt, data.spectral, data.opts,
                 integrate.layer = "angles_layers", k.maxit = 100,
                 get.plots = FALSE)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol,  Joris Timmermans (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
