% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ephoton.R
\name{get.ephoton}
\alias{get.ephoton}
\title{get.ephoton}
\usage{
get.ephoton(lambda, constants)
}
\arguments{
\item{lambda}{numeric vector. Wavelength(s) of the photon(s) (m).}

\item{constants}{data.frame. Physical constants table (as used throughout SCOPEinR), must contain rows for \code{h} (Planck's constant, J s) and \code{c} (speed of light, m s^-1) in columns \code{constant}/\code{value}.}
}
\value{
numeric vector. Energy content \code{E} of one photon at each wavelength \code{lambda} (J).
}
\description{
\code{get.ephoton} calculates the energy content (J) of 1 photon of wavelength lambda (m)
}
\examples{
\dontrun{
E <- get.ephoton(lambda, constants)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Wout Verhoef, Christiaan van der Tol, Joris Timmermans (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
