% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{get.Fluorescence.model}
\alias{get.Fluorescence.model}
\title{get.Fluorescence.model
\code{get.Fluorescence.model} Fluorescence model}
\usage{
get.Fluorescence.model(ps, x, Kp, Kf, Kd, Knparams)
}
\arguments{
\item{ps}{numeric. Photochemical yield (fraction of absorbed light used in photochemistry).}

\item{x}{numeric. Normalized (0-1) intermediate variable driving non-photochemical quenching, passed in to avoid recomputing it internally.}

\item{Kp}{numeric. Rate constant for photochemistry.}

\item{Kf}{numeric. Rate constant for fluorescence.}

\item{Kd}{numeric. Rate constant for (light-independent) thermal dissipation.}

\item{Knparams}{numeric vector of length 3: Kno (max NPQ rate constant), alpha, beta - shape parameters of the non-photochemical quenching response.}
}
\value{
A list with eta, qE, qQ, fs, fo, fm, fo0, fm0, and Kn - fluorescence yields and quenching parameters.
}
\description{
get.Fluorescence.model
\code{get.Fluorescence.model} Fluorescence model
}
\examples{
get.Fluorescence.model(ps = 0.3, x = 0.5, Kp = 4, Kf = 0.05, Kd = 0.95,
                        Knparams = c(2.48, 2.83, 0.114))

}
