% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{get.gsFun}
\alias{get.gsFun}
\title{get.gsFun
\code{get.gsFun}   get stomatal conductance}
\usage{
get.gsFun(Cs, RH, A, BallBerrySlope, BallBerry0)
}
\arguments{
\item{Cs}{numeric. CO2 concentration at the leaf surface.}

\item{RH}{numeric. Relative humidity (0-1).}

\item{A}{numeric. Net assimilation rate, in the same CO2 units as Cs, per m2 per s.}

\item{BallBerrySlope}{numeric. Slope parameter of the Ball-Berry stomatal conductance model.}

\item{BallBerry0}{numeric. Intercept (minimum conductance) parameter of the Ball-Berry model.}
}
\value{
gs: stomatal conductance.
}
\description{
get.gsFun
\code{get.gsFun}   get stomatal conductance
}
\examples{
get.gsFun(Cs = 400, RH = 0.7, A = 15, BallBerrySlope = 9, BallBerry0 = 0.01)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
