% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatfluxes.R
\name{get.heatfluxes}
\alias{get.heatfluxes}
\title{get.heatfluxes
\code{get.heatfluxes} Calculates  latent and sensible heat flux}
\usage{
get.heatfluxes(ra, rs, Tc, ea, Ta, e_to_q, Ca, Ci)
}
\arguments{
\item{ra}{aerodynamic resistance for heat         s m-1}

\item{rs}{stomatal resistance                     s m-1}

\item{Tc}{leaf temperature                        oC}

\item{ea}{vapour pressure above canopy            hPa}

\item{Ta}{air temperature above canopy            oC}

\item{e_to_q}{conv. from vapour pressure to abs hum   hPa-1}

\item{Ca}{ambient CO2 concentration               umol m-3}

\item{Ci}{intercellular CO2 concentration         umol m-3}
}
\value{
a list with
lEc         latent heat flux of a leaf              W m-2
}
\description{
get.heatfluxes
\code{get.heatfluxes} Calculates  latent and sensible heat flux
}
\examples{
\dontrun{
out <- get.heatfluxes(ra, rs, Tc, ea, Ta, e_to_q, Ca, Ci)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
