% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{get.high.temp.inhibtionC3}
\alias{get.high.temp.inhibtionC3}
\title{get.high.temp.inhibtionC3
\code{get.high.temp.inhibtionC3} High Temperature Inhibition Function:The following function pertains to C3 photosynthesis}
\usage{
get.high.temp.inhibtionC3(Tref, R, T, deltaS, deltaHd)
}
\arguments{
\item{Tref}{numeric. Reference temperature, in Kelvin.}

\item{R}{numeric. Universal gas constant.}

\item{T}{numeric. Leaf temperature, in Kelvin.}

\item{deltaS}{numeric. Entropy term of the high-temperature inhibition function.}

\item{deltaHd}{numeric. Deactivation energy (enthalpy of high-temperature inhibition).}
}
\value{
fHTv: high-temperature inhibition factor applied to C3 photosynthesis rates.
}
\description{
get.high.temp.inhibtionC3
\code{get.high.temp.inhibtionC3} High Temperature Inhibition Function:The following function pertains to C3 photosynthesis
}
\examples{
get.high.temp.inhibtionC3(Tref = 298.15, R = 8.314, T = 305, deltaS = 650, deltaHd = 200000)
}
