% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{get.MD12}
\alias{get.MD12}
\title{MD12 algorithm for the computation of fluorescence yield}
\usage{
get.MD12(ps, Ja, Jms, kps, kf, kds, kDs)
}
\arguments{
\item{ps}{numeric. Photochemical yield (fraction of absorbed light used in photochemistry).}

\item{Ja}{numeric. Actual electron transport rate.}

\item{Jms}{numeric. Maximum (light-saturated) electron transport rate.}

\item{kps}{numeric. Rate constant for photochemistry.}

\item{kf}{numeric. Rate constant for fluorescence.}

\item{kds}{numeric. Rate constant for (light-independent) thermal dissipation.}

\item{kDs}{numeric. Rate constant for (light-dependent, sustained) thermal dissipation.}
}
\value{
fs: PSII fluorescence yield.
}
\description{
MD12 algorithm for the computation of fluorescence yield
}
\examples{
get.MD12(ps = 0.3, Ja = 40, Jms = 120, kps = 4, kf = 0.05, kds = 0.95, kDs = 0)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
