% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluspect_functions.R
\name{get.numjacobian}
\alias{get.numjacobian}
\title{Compute the numerical Jacobian of the Fluspect leaf model with respect to its parameters}
\usage{
get.numjacobian(x, spectral, inputLeaf, optipar)
}
\arguments{
\item{x}{numeric vector. Leaf biochemistry/structure parameters at which to compute the numerical Jacobian.}

\item{spectral}{list. Spectral configuration used by the Fluspect model.}

\item{inputLeaf}{list. Leaf input parameters.}

\item{optipar}{list. Optical parameters used by the Fluspect model.}
}
\value{
J: a numeric array of partial derivatives (reflectance/transmittance with respect to each parameter in x).
}
\description{
Compute the numerical Jacobian of the Fluspect leaf model with respect to its parameters
}
\examples{
\dontrun{
get.numjacobian(x, spectral, inputLeaf, optipar)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Wout Verhoef, Christiaan van der Tol, Joris Timmermans (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
