% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.outs.in.R
\name{get.outs.in}
\alias{get.outs.in}
\title{\code{get.outs.in} get the outputs from the SCOPE model. This code is for single simulations}
\usage{
get.outs.in(data.sim, LUT, path.out, get.plots, get.more.inputs)
}
\arguments{
\item{data.sim}{a list elements generated by the SCOPE model}

\item{LUT}{a list of inputs used by the SCOPE model}

\item{path.out}{path for saving the main files}

\item{get.plots}{is true, intermediate plots will be generated in Plot folder}

\item{get.more.inputs}{is true, intermediate plots will be generated in Additional.inputs folder}
}
\description{
\code{get.outs.in} get the outputs from the SCOPE model. This code is for single simulations
}
\examples{
\dontrun{
get.outs.in(data.sim, LUT, path.out, get.plots = TRUE, get.more.inputs = FALSE)
}
}
\author{
Carlos Camino (Ported version into R)
}
