% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.outs.lut.R
\name{get.outs.lut}
\alias{get.outs.lut}
\title{\code{get.outs.lut} get the outputs from SCOPE model. This code is for a LUT of simulations}
\usage{
get.outs.lut(data.sim, LUT, path.out, get.plots, nSamples)
}
\arguments{
\item{data.sim}{list elements generated by the SCOPE model}

\item{LUT}{a list of inputs used in the SCOPE model}

\item{path.out}{path for saving the main files}

\item{get.plots}{is true, intermediate plots will be generated in Plot folder}

\item{nSamples}{integer. Number of simulations (rows of \code{LUT}/\code{data.sim}) to process.}
}
\description{
\code{get.outs.lut} get the outputs from SCOPE model. This code is for a LUT of simulations
}
\examples{
\dontrun{
get.outs.lut(data.sim, LUT, path.out, get.plots = TRUE, nSamples = 100)
}
}
\author{
Carlos Camino (Ported version into R)
}
