% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.outs.lut.v2.R
\name{get.outs.lut.v2}
\alias{get.outs.lut.v2}
\title{\code{get.outs.lut} get the outputs from SCOPE model. This code is for a LUT of simulations}
\usage{
get.outs.lut.v2(data.sim, LUT, path.out, get.plots, get.outputs, nSamples)
}
\arguments{
\item{data.sim}{list elements generated by the SCOPE model}

\item{LUT}{a list of inputs used in the SCOPE model}

\item{path.out}{path for saving the main files}

\item{get.plots}{is true, intermediate plots will be generated in Plot folder}

\item{get.outputs}{If get.outputs = 'ALL', all variables will be retrieved; if get.outputs = 'Main', only the main variables will be retrieved.
By default, SCOPE uses get.outputs = 'ALL', for processing a huge LUT, it is recommended to use get.outputs = 'Main'.}

\item{nSamples}{number of LUT samples to process; used to size progress reporting for the extraction loop.}
}
\description{
\code{get.outs.lut} get the outputs from SCOPE model. This code is for a LUT of simulations
}
\examples{
\dontrun{
nSamples <- 100
get.outs.lut.v2(data.sim = LUT, LUT = LUT, path.out = path.out, get.plots = TRUE,
                 get.outputs = 'ALL', nSamples = nSamples)
}
}
\author{
Carlos Camino (Ported version into R)
}
