% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Planck.R
\name{get.Planck}
\alias{get.Planck}
\title{Planck function}
\usage{
get.Planck(wl, Tb, em = NULL)
}
\arguments{
\item{wl}{numeric vector. Wavelength(s) of interest, in nanometers (nm).}

\item{Tb}{numeric vector. Temperature(s) of the emitting object, in Kelvin (K).}

\item{em}{numeric vector, optional. Emissivity of the object at each wavelength/temperature (dimensionless, 0-1). If missing, emissivity is set to 1 (ideal blackbody) for every element of \code{Tb}.}
}
\value{
numeric vector. Spectral radiance \code{Lb} emitted at each wavelength/temperature pair (W m^-2 sr^-1 um^-1).
}
\description{
\code{get.Planck} estimates the spectral radiance emitted by a blackbody
(or greybody, if \code{em} is supplied) at given wavelength(s) and
temperature(s), following Planck's radiation law.
}
\examples{
\dontrun{
Lb <- get.Planck(wl, Tb)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
