% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistances.R
\name{get.psih}
\alias{get.psih}
\title{subfunction ph for stability correction (eg. Paulson, 1970)}
\usage{
get.psih(z, L, unstable, stable, x)
}
\arguments{
\item{z}{numeric. Height above the surface, in meters.}

\item{L}{numeric. Monin-Obukhov length, in meters (atmospheric stability scale).}

\item{unstable}{logical. TRUE if the atmosphere is in an unstable stratification regime.}

\item{stable}{logical. TRUE if the atmosphere is in a stable stratification regime.}

\item{x}{numeric. Stability correction intermediate variable, typically (1 - 16*z/L)^0.25.}
}
\value{
ph vector
}
\description{
subfunction ph for stability correction (eg. Paulson, 1970)
}
\examples{
get.psih(z = 2, L = -50, unstable = TRUE, stable = FALSE, x = 1.2)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
