% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMf.R
\name{get.RTMf}
\alias{get.RTMf}
\title{get.RTMf
\code{get.RTMf} Calculates the spectrum of fluorescent radiance in the observer's direction and also
the TOC spectral hemispherical upward Fs flux.}
\usage{
get.RTMf(
  data.spectral,
  data.rad,
  data.soil,
  data.leafopt,
  data.canopy,
  data.gap,
  data.angles,
  data.etau,
  data.etah,
  get.plots = T
)
}
\arguments{
\item{data.spectral}{information about wavelengths and resolutions}

\item{data.rad}{a large number of radiative fluxes: spectrally distributed and integrated, and canopy radiative transfer coefficients.}

\item{data.soil}{soil properties}

\item{data.leafopt}{leaf optical properties}

\item{data.canopy}{canopy properties (such as LAI and height)}

\item{data.gap}{probabilities of direct light penetration and viewing}

\item{data.angles}{viewing and observation angles}

\item{data.etau}{relative fluorescence emission efficiency for sunlit leaves}

\item{data.etah}{relative fluorescence emission efficiency for shaded leaves}

\item{get.plots}{logical. If \code{TRUE}, intermediate diagnostic plots are produced. Default \code{TRUE}.}
}
\value{
a rad object with a large number of radiative fluxes: spectrally distributed
and integrated, and canopy radiative transfer coefficients.
}
\description{
get.RTMf
\code{get.RTMf} Calculates the spectrum of fluorescent radiance in the observer's direction and also
the TOC spectral hemispherical upward Fs flux.
}
\details{
Date:  12 Dec 2007.
Update history:
\itemize{
\item 26 Aug 2008 CvdT        small correction to matrices
\item 07 Nov 2008 CvdT        changed layout
\item 19 Mar 2009 CvdT        major corrections: lines 95-96,101-107, and 119-120.
\item 07 Apr 2009 WV & CvdT   major correction: lines 89-90, azimuth
dependence was not there in previous verions (implicit assumption of  azimuth(solar-viewing) = 0). This has been corrected
\item May-June 2012 WV & CvdT Add calculation of hemispherical Fs fluxes
\item Jan-Feb 2013 WV         Inputs and outputs via structures for SCOPE Version 1.40
\item Aug-Oct 2016 PY         Re-write the calculation of emitted SIF of each layer. It doesnt use loop at all. with the function bsxfun, the  calculation is much faster
\item Oct 2017-Feb 2018 PY    Re-write the RTM of fluorescence
\item Jan 2020 CvdT           Modified to include 'lite' option, mSCOPE representation
\item 25 Jun 2020 PY          Po, Ps, Pso. fix the problem we have with the oblique angles above 80 degrees
}
}
\examples{
\dontrun{
get.RTMf(data.spectral, data.rad, data.soil, data.leafopt, data.canopy,
         data.gap, data.angles, data.etau, data.etah)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Wout Verhoef and Christiaan van der Tol  (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
