% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMt_planck.R
\name{get.RTMt.planck}
\alias{get.RTMt.planck}
\title{get.RTMt.planck
\code{get.RTMt.planck}analogue to get.RTMt.sb, this function calculates total outgoing radiation in hemispherical
direction and total absorbed radiation per leaf and soil component.
Radiation is integrated over the whole thermal spectrum with Stefan-Boltzman's equation. This function is a simplified version of
'get.RTMt.planck', and is less time consuming since it does not do the calculation for each wavelength separately.}
\usage{
get.RTMt.planck(
  data.spectral,
  data.rad,
  data.soil,
  data.leafbio,
  data.leafopt,
  data.canopy,
  data.gap,
  Tcu,
  Tch,
  Tsu,
  Tsh,
  get.plots = T
)
}
\arguments{
\item{data.spectral}{information about wavelengths and resolutions}

\item{data.rad}{a large number of radiative fluxes: spectrally distributed and integrated, and canopy radiative transfer coefficients.}

\item{data.soil}{soil properties}

\item{data.leafbio}{leaf properties (Cab....)}

\item{data.leafopt}{leaf optical properties}

\item{data.canopy}{canopy properties (such as LAI and height)}

\item{data.gap}{probabilities of direct light penetration and viewing}

\item{Tcu}{Temperature of sunlit leaves    (oC), (13x36x60)}

\item{Tch}{Temperature of shaded leaves    (oC), (13x36x60)}

\item{Tsu}{Temperature of sunlit soil      (oC), (1)}

\item{Tsh}{Temperature of shaded soil      (oC), (1)}

\item{get.plots}{is true plot the intermediate plots}
}
\value{
a large number of radiative fluxes: spectrally distributed and integrated, and canopy radiative transfer coefficients., Here
thermal fluxes are added.
}
\description{
Date:   05  Nov 2007
Update:
\itemize{
\item 13 Nov 2007
\item 16 Nov 2007 CvdT: improved calculation of net radiation.
\item 27 Mar 2008 JT: added directional calculation of radiation.
\item 24 Apr 2008 JT: introduced dx as thickness of layer (see parameters).
\item 31 Oct 2008 JT: introduced optional directional calculation.
\item 31 Oct 2008 JT: changed initialisation of F1 and F2 -> zeros.
\item 07 Nov 2008 CvdT: changed layout.
\item 16 Mar 2009 CvdT: removed Tbright calculation.
\item Feb 2013 WV: introduces structures for version 1.40.
\item 04 Dec 2019 CvdT: adapted for SCOPE-lite.
\item 17 Mar 2020 CvdT: mSCOPE representation.
}
}
\examples{
\dontrun{
data.rad <- get.RTMt.planck(data.spectral, data.rad, data.soil, data.leafbio,
                             data.leafopt, data.canopy, data.gap,
                             Tcu, Tch, Tsu, Tsh, get.plots = FALSE)
}
}
\author{
Wout Verhoef and Christiaan van der Tol  (Original version in Matlab)

Carlos Camino (Ported version into R)
}
