% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMz.R
\name{get.RTMz}
\alias{get.RTMz}
\title{get.RTMz
\code{get.RTMz} Calculates the small modification of TOC outgoing radiance
due to the conversion of Violaxanthin into Zeaxanthin in leaves}
\usage{
get.RTMz(
  data.spectral,
  data.rad,
  data.soil,
  data.leafopt,
  data.canopy,
  data.gap,
  data.angles,
  data.Knu,
  data.Knh,
  get.plots = T
)
}
\arguments{
\item{data.spectral}{information about wavelengths and resolutions}

\item{data.rad}{a large number of radiative fluxes: spectrally distributed and integrated, and canopy radiative transfer coefficients.}

\item{data.soil}{soil properties}

\item{data.leafopt}{leaf optical properties}

\item{data.canopy}{canopy properties (such as LAI and height)}

\item{data.gap}{probabilities of direct light penetration and viewing}

\item{data.angles}{viewing and observation angles}

\item{data.Knu}{...kn for  sunlit leaves (data.bcu$Kn)}

\item{data.Knh}{.. kn for shaded leavess (data.bch$Kn)}

\item{get.plots}{is true plot the intermediate plots}
}
\value{
a rad object with a large number of radiative fluxes: spectrally distributed
and integrated, and canopy radiative transfer coefficients.
}
\description{
Date:   08 Dec 2016
Update:
\itemize{
\item 17 Mar 2020 CvdT: added cluming, mSCOPE representation
\item 25 Jun 2020 CvdT: Po, Ps, Pso. fix the problem we have with the oblique angles above 80 degrees
}
}
\examples{
\dontrun{
data.rad <- get.RTMz(data.spectral, data.rad, data.soil, data.leafopt,
                      data.canopy, data.gap, data.angles, data.Knu, data.Knh,
                      get.plots = FALSE)
}
}
\author{
Christiaan van der Tol  (Original version in Matlab)

Carlos Camino (Ported version into R)
}
