% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.SCOPE.outputs.R
\name{get.SCOPE.outputs}
\alias{get.SCOPE.outputs}
\title{\code{get.SCOPE.outputs} get the outputs from SCOPE model}
\usage{
get.SCOPE.outputs(
  data.sim,
  N.sims,
  LUT,
  path.out,
  get.plots,
  get.outputs = NULL,
  get.more.inputs
)
}
\arguments{
\item{data.sim}{a list elements generated by the SCOPE model}

\item{N.sims}{a vector with number of simulations generated by the SCOPE model}

\item{LUT}{a LUT with the inputs used in the SCOPE model}

\item{path.out}{path for saving the main files}

\item{get.plots}{is true, intermediate plots will be generated in Plot folder}

\item{get.outputs}{If get.outputs = 'ALL', all variables will be retrieved; if get.outputs = 'Main', only the main variables will be retrieved.}

\item{get.more.inputs}{is true, intermediate plots will be generated in Additional.inputs folder}
}
\description{
\code{get.SCOPE.outputs} get the outputs from SCOPE model
}
\examples{
\dontrun{
out <- get.SCOPE.outputs(data.sim, N.sims, LUT, path.out, get.plots = FALSE,
                          get.outputs = "Main", get.more.inputs = FALSE)
}
}
\author{
Carlos Camino (Ported version into R)
}
