% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.SCOPE.parallel.R
\name{get.SCOPE.parallel}
\alias{get.SCOPE.parallel}
\title{Run SCOPE simulations in parallell}
\usage{
get.SCOPE.parallel(
  LUT,
  options.SCOPE = NULL,
  optipar = NULL,
  path.out = NULL,
  parallel = TRUE,
  canopy.model = "fourSAIL",
  leaf.model = "fluspect-CX",
  get.outputs = "Main",
  get.plots = TRUE,
  get.csv = T,
  n.cores = 8
)
}
\arguments{
\item{LUT}{Leaf, biochemistry, viewing angles, meteo, and canopy properties needed for running SCOPE model.}

\item{options.SCOPE}{Optical leaf properties, the total irradiance if a specific value is provided instead of the usual Modtran output.}

\item{optipar}{list. Leaf-level optical parameters passed on to \code{\link{get.SCOPE}} for each LUT chunk (e.g. \code{SCOPEinR::optipar2021.Pro.CX}).}

\item{path.out}{Folder for saving the SCOPE outputs.}

\item{parallel}{Logical, indicating whether to use parallel processing. Default is TRUE.}

\item{canopy.model}{Selection of canopy model options available are 'fourSAIL', 'fourSAIL' and 'INFORM'. By default, the fourSAIL model will be used.}

\item{leaf.model}{Selection of canopy model options available are 'fluspect-CX', 'fluspect-B', 'PROSPECT', 'Liberty'. By default, the fluspect-CX model will be used.}

\item{get.outputs}{If get.outputs = 'ALL', all variables will be retrieved; if get.outputs = 'Main', only the main variables will be retrieved.
By default, SCOPE uses get.outputs = 'ALL', for processing a huge LUT, it is recommended to use get.outputs = 'Main'.}

\item{get.plots}{Logical, indicating whether to plot the intermediate plots. Default is TRUE.}

\item{get.csv}{logical, indicating whether to save outputs. Default is TRUE.}

\item{n.cores}{Integer, indicating the number of cores to use. if this parameter is null or missing Detect cores - 2 will be used}
}
\value{
A list of SCOPE simulations.
}
\description{
Run SCOPE simulations in parallell
}
\examples{
\dontrun{
get.SCOPE.parallel(LUT, options.SCOPE = NULL, optipar = NULL,
                   path.out = "output", parallel = TRUE, leaf.model = 'fluspect-CX',
                   canopy.model = 'fourSAIL', get.outputs = 'Main', get.plots = TRUE)
}

}
