% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.SCOPE.plots.R
\name{get.SCOPE.plots}
\alias{get.SCOPE.plots}
\title{\code{get.SCOPE.plots} get simulations based on SCOPE model}
\usage{
get.SCOPE.plots(path.files, plant.trait, get.plots)
}
\arguments{
\item{path.files}{folder with main tables from  the SCOPE simulation.}

\item{plant.trait}{plant trait}

\item{get.plots}{type plot for comparing, options are: 'fluorescence', 'reflectance', 'radiance'}
}
\description{
\code{get.SCOPE.plots} get simulations based on SCOPE model
}
\examples{
\dontrun{
get.SCOPE.plots(path.files = "outs", plant.trait = "Cab", get.plots = "reflectance")
}
}
\author{
Carlos Camino (Ported version into R)
}
