% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_bands.R
\name{get.spectra.SCOPE}
\alias{get.spectra.SCOPE}
\title{get.spectra.SCOPE
\code{get.spectra.SCOPE} Calculates the spectra of hemisperical and directional observed #' a function to get the spectral characteristics for SCOPE model}
\usage{
get.spectra.SCOPE(getSpectral = T)
}
\arguments{
\item{getSpectral}{is T get the spectral object needed for SCOPE}
}
\value{
a list with al spectral ranges and Intervals
}
\description{
get.spectra.SCOPE
\code{get.spectra.SCOPE} Calculates the spectra of hemisperical and directional observed #' a function to get the spectral characteristics for SCOPE model
}
\examples{
spectral <- get.spectra.SCOPE(getSpectral = TRUE)
}
