% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Stefan_Boltzmann.R
\name{get.Stefan_Boltzmann}
\alias{get.Stefan_Boltzmann}
\title{Stefan-Boltzmann equation}
\usage{
get.Stefan_Boltzmann(T_C)
}
\arguments{
\item{T_C}{numeric vector. Temperature(s) of the emitting surface, in degrees Celsius.}
}
\value{
numeric vector. Blackbody radiant exitance \code{H} (W m^-2) at each temperature in \code{T_C}.
}
\description{
\code{get.Stefan_Boltzmann} computes the total radiant exitance of a
blackbody at a given temperature, using the Stefan-Boltzmann law
(\eqn{H = \sigma T^4}).
}
\examples{
\dontrun{
H <- get.Stefan_Boltzmann(T_C = 20)
}
}
\author{
Carlos Camino
}
