% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{get.temperature.functionC3}
\alias{get.temperature.functionC3}
\title{get.temperature.functionC3
\code{get.temperature.functionC3} Temperature Correction Functions:The following function pertains to C3 photosynthesis}
\usage{
get.temperature.functionC3(Tref, R, Temp, deltaHa)
}
\arguments{
\item{Tref}{numeric. Reference temperature, in Kelvin.}

\item{R}{numeric. Universal gas constant.}

\item{Temp}{numeric. Leaf temperature, in Kelvin.}

\item{deltaHa}{numeric. Activation energy of the temperature-dependent process.}
}
\value{
fTv: Arrhenius-type temperature correction factor.
}
\description{
get.temperature.functionC3
\code{get.temperature.functionC3} Temperature Correction Functions:The following function pertains to C3 photosynthesis
}
\examples{
get.temperature.functionC3(Tref = 298.15, R = 8.314, Temp = 305, deltaHa = 65330)
}
