% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMo_functions.R
\name{get.volscatt.scope}
\alias{get.volscatt.scope}
\title{get.volscatt.scope version 2.0 from SCOPE model}
\usage{
get.volscatt.scope(tts, tto, psi, ttli)
}
\arguments{
\item{tts}{Sun: zenith angle in degrees}

\item{tto}{observation:zenith angle in degrees}

\item{psi}{Difference of  azimuth angle between solar and viewing position}

\item{ttli}{leaf inclination array}
}
\value{
The function returns a list of four items: "chi_s", "chi_o", "frho", and "ftau".
These items represent the scattering phase functions for direct and diffuse radiation, respectively.
}
\description{
\code{volscatt.scope} calculates the scattering phase functions using a
radiative transfer model based on the optical properties of leaves. Specifically,
the model computes the fraction of radiation scattered in the forward and backward directions
for a given set of input parameters
}
\details{
Date: 11 February 2008.
}
\examples{
get.volscatt.scope(tts = 30, tto = 10, psi = 0, ttli = seq(5, 85, by = 10))
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Wout Verhoef, Joris Timmermans (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
