% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{getBSM}
\alias{getBSM}
\title{Brightness-Shape-Moisture soil model}
\usage{
getBSM(soilpar, spec, emp)
}
\arguments{
\item{soilpar}{list. Soil parameters: \code{BSMBrightness} (overall soil brightness), \code{BSMlat} (spectral shape "latitude", typical range 20-40 deg), \code{BSMlon} (spectral shape "longitude", typical range 45-65 deg).}

\item{spec}{list. Spectral inputs: \code{GSV} (Global Soil Vectors basis spectra, matrix nwl x 3), \code{Kw} (water absorption coefficient spectrum, nwl x 1), \code{nw} (refraction index of water spectrum, nwl x 1).}

\item{emp}{list. Empirical wetting parameters: \code{SMp} (soil moisture volume percentage, range 5-55), \code{SMC} (soil moisture capacity, recommended 0.25), \code{film} (effective optical thickness of a single water film, recommended 0.015).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Wet soil reflectance spectra across 400 nm to 2400 nm
}\if{html}{\out{</div>}}
}
\description{
\code{getBSM} Calculates the spectrum of fluorescent radiance in the observer's direction and also
the TOC spectral hemispherical upward Fs flux.
}
\examples{
\dontrun{
save.soil <- getBSM(soilpar, spec, emp)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
