% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluspect_B_ForSCOPE.R
\name{getFluspect.B.SCOPE}
\alias{getFluspect.B.SCOPE}
\title{Leaf FLUSPECT-B model for SCOPE}
\usage{
getFluspect.B.SCOPE(inputsLeaf, inputsOptipar, version = "D")
}
\arguments{
\item{inputsLeaf}{a LUT for FLUXspect}

\item{inputsOptipar}{internal parameters}

\item{version}{Leaf model uses for Kcal: 'D' refers as PROSPECT-D; 'PRO', refers as PROSPECT-PRO}
}
\value{
a list which contains:
\itemize{
\item refl (reflectance)
\item tran (transmittance)
\item Mb (backward scattering fluorescence matrix, I for PSI and II for PSII)
\item Mf (forward scattering fluorescence matrix,  I for PSI and II for PSII)
}
}
\description{
\code{getFluspect.B.SCOPE} calculates reflectance and transmittance spectra of a leaf using FLUSPECT-B,
calculates reflectance and transmittance spectra of a leaf using FLUSPECT-B,
plus four excitation-fluorescence matrices
}
\examples{
\dontrun{
inputs <- ToolsRTM::inputsRTMs
LUT <- as.data.frame(ToolsRTM::getLUT(inputs = inputs, nLUT = 1, setseed = 1234))
sim <- getFluspect.B.SCOPE(inputsLeaf = LUT, inputsOptipar = SCOPEinR::optipar, version = 'D')
}
}
\author{
Wout Verhoef, Christiaan van der Tol, Joris Timmermans, Nastassia Vilfan (Original version in Matlab)

Carlos Camino (Ported version into R)
}
