% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLUT.SCOPE-v1.R
\name{getLUT.SCOPE.v1}
\alias{getLUT.SCOPE.v1}
\title{Generate LUT for SCOPE (v1)}
\usage{
getLUT.SCOPE.v1(inputLUT, nLUT = 100, setseed = 123)
}
\arguments{
\item{inputLUT}{A table (LUT) containing data.}

\item{nLUT}{Number of samples to generate from the input LUT. Defaults to 100.}

\item{setseed}{Seed value to ensure reproducibility of results. Defaults to 123.}
}
\value{
Returns a data frame containing the generated LUT samples.
}
\description{
This function generates samples of a Look-Up Table (LUT) to analyze variations
in plant traits required for a specific scope.
}
\examples{
\dontrun{
inputLUT <- read.table(system.file("input", "inputs_SCOPE.csv", package = "SCOPEinR"),
                       header = TRUE, sep = ",")
getLUT.SCOPE.v1(inputLUT = inputLUT, nLUT = 100, setseed = 123)
}
}
