% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MD12.R
\name{MD12}
\alias{MD12}
\title{MD12 algorithm for the computation of fluorescence yield}
\usage{
MD12(ps, Ja, Jms, kps, kf, kds, kDs)
}
\arguments{
\item{ps}{numeric vector. PSII photochemical yield (probability that an absorbed photon leads to charge separation), dimensionless [E/E], range 0-1.}

\item{Ja}{numeric vector. Actual (CO2-limited) electron transport rate, in the same units as \code{Jms} (typically umol electrons m^-2 s^-1).}

\item{Jms}{numeric vector. Potential (light-limited) electron transport rate, reduced for PSII photodamage, in the same units as \code{Ja}.}

\item{kps}{numeric vector. Rate constant for photochemistry at PSII (steady state), dimensionless relative rate constant.}

\item{kf}{numeric vector. Rate constant for fluorescence emission, dimensionless relative rate constant.}

\item{kds}{numeric vector. Rate constant for constitutive (basal, dark-type) thermal dissipation at PSII, steady state, dimensionless relative rate constant.}

\item{kDs}{numeric vector. Rate constant for regulated (NPQ-type) thermal dissipation at PSII, steady state, dimensionless relative rate constant.}
}
\value{
numeric vector. PSII fluorescence quantum yield \code{fs}, dimensionless [E/E], computed as the minimum of the CO2-limited and light-limited yields.
}
\description{
\code{MD12} computes the PSII fluorescence quantum yield from the electron
transport rate and a set of first-order rate constants describing the
competing de-excitation pathways of PSII (photochemistry, fluorescence,
constitutive and regulated thermal dissipation). The yield is evaluated
both under CO2-limited and under light-limited conditions, and the
smaller (limiting) value is returned, following the model described in
Van der Tol et al. (2014, Biogeosciences) building on Magnani et al. (2012).
}
\examples{
\dontrun{
fs <- MD12(ps, Ja, Jms, kps, kf, kds, kDs)
}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
