% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo-timeseries}
\alias{meteo-timeseries}
\alias{Rin_}
\alias{Rli_}
\alias{Esun_}
\alias{Esky_}
\alias{Ta_}
\alias{ea_}
\alias{p_}
\alias{u_}
\alias{t_}
\alias{year_}
\title{Example meteorological/radiation time series for diurnal SCOPE simulations}
\format{
A numeric vector or data frame, one row/element per time step.

An object of class \code{data.frame} with 214 rows and 1 columns.

An object of class \code{data.frame} with 2162 rows and 1 columns.

An object of class \code{data.frame} with 2162 rows and 1 columns.

An object of class \code{data.frame} with 214 rows and 1 columns.

An object of class \code{data.frame} with 214 rows and 1 columns.

An object of class \code{data.frame} with 214 rows and 1 columns.

An object of class \code{data.frame} with 214 rows and 1 columns.

An object of class \code{data.frame} with 214 rows and 1 columns.

An object of class \code{data.frame} with 216 rows and 1 columns.
}
\usage{
Rin_

Rli_

Esun_

Esky_

Ta_

ea_

p_

u_

t_

year_
}
\description{
Example time series (one value per time step) used by \code{\link{getLUT_time}}
to drive multi-timestep SCOPE runs: incoming shortwave radiation
(\code{Rin_}), incoming longwave radiation (\code{Rli_}), sun/sky
irradiance spectra (\code{Esun_}/\code{Esky_}), air temperature
(\code{Ta_}), vapour pressure (\code{ea_}), air pressure (\code{p_}),
wind speed (\code{u_}), time-of-day (\code{t_}) and year (\code{year_}).
}
\keyword{datasets}
