% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optipar-datasets}
\alias{optipar-datasets}
\alias{optipar}
\alias{optipar.2015}
\alias{optipar2017.ProspectD}
\alias{optipar2020.prospectD.BSM2019}
\alias{optipar2021.Pro.CX}
\title{Leaf optical parameters (PROSPECT/Fluspect, various parameterizations)}
\format{
A data frame/list with one row per wavelength and one column per optical coefficient.

An object of class \code{data.frame} with 2001 rows and 15 columns.

An object of class \code{list} of length 15.

An object of class \code{list} of length 16.

An object of class \code{list} of length 18.
}
\source{
\code{optipar}, \code{optipar.2015}: Feret et al. 2008 (PROSPECT-4); \code{optipar2017.ProspectD}: Feret et al. 2017 (PROSPECT-D);
\code{optipar2020.prospectD.BSM2019}: Feret et al. 2017 + BSM soil model (Yang et al. 2019); \code{optipar2021.Pro.CX}: PROSPECT-PRO (Feret et al. 2021).
}
\usage{
optipar

optipar.2015

optipar2017.ProspectD

optipar2020.prospectD.BSM2019

optipar2021.Pro.CX
}
\description{
Wavelength-dependent leaf-level optical coefficients (refractive index,
specific absorption coefficients per pigment, PSI/PSII quantum yield
spectra) required by the Fluspect/PROSPECT leaf models. Several versions
are bundled, corresponding to different published parameterizations.
}
\keyword{datasets}
