% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satvap.R
\name{satvap}
\alias{satvap}
\title{calculates the saturated vapour pressure at temperature T (degrees C)
and the derivative of es to temperature s (kPa/C)}
\usage{
satvap(Temp)
}
\arguments{
\item{Temp}{numeric. Air/leaf temperature, in degrees Celsius.}
}
\value{
es: the saturated vapour pressure at Temp, in hPa/mbar.
}
\description{
calculates the saturated vapour pressure at temperature T (degrees C)
and the derivative of es to temperature s (kPa/C)
}
\details{
Date: 2003.
}
\examples{
satvap(20)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
