% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scope-input-datasets}
\alias{scope-input-datasets}
\alias{inputsSCOPE}
\alias{input_border}
\title{Example SCOPE input LUT and input border/range definitions}
\format{
A data frame, one row per input parameter, with default value and (for \code{input_border}) min/max range columns.

An object of class \code{data.frame} with 26 rows and 7 columns.
}
\usage{
inputsSCOPE

input_border
}
\description{
\code{inputsSCOPE} is an example base look-up table of SCOPE input
parameters (as read from \code{inst/input/inputs_SCOPE.csv}, used by
\code{\link{getLUT.SCOPE}} to generate randomized LUTs). \code{input_border}
gives the valid minimum/maximum range for each input parameter.
}
\keyword{datasets}
