% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biochemical_functions.R
\name{sel_root}
\alias{sel_root}
\title{quadratic formula, root of least magnitude}
\usage{
sel_root(a, b, c, dsign)
}
\arguments{
\item{a}{numeric. Quadratic coefficient (a in a\emph{x^2 + b}x + c = 0).}

\item{b}{numeric. Linear coefficient.}

\item{c}{numeric. Constant term.}

\item{dsign}{numeric. Sign of the discriminant term to select which root to return (+1 or -1).}
}
\value{
The selected root of least magnitude.
}
\description{
quadratic formula, root of least magnitude
}
\examples{
sel_root(a = 1, b = -3, c = 2, dsign = 1)
}
