% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_satvap.R
\name{slope_satvap}
\alias{slope_satvap}
\title{calculates the saturated vapour pressure at temperature T (degrees C)
and the derivative of es to temperature s (kPa/C)}
\usage{
slope_satvap(Temp)
}
\arguments{
\item{Temp}{numeric. Air/leaf temperature, in degrees Celsius.}
}
\value{
The derivative of saturated vapour pressure with respect to temperature (slope), in kPa/degree C.
}
\description{
calculates the saturated vapour pressure at temperature T (degrees C)
and the derivative of es to temperature s (kPa/C)
}
\details{
Date: 2003.
}
\examples{
slope_satvap(20)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
