% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_respiration.R
\name{soil_respiration}
\alias{soil_respiration}
\title{soil respiration}
\usage{
soil_respiration(Ts)
}
\arguments{
\item{Ts}{numeric. Soil temperature.}
}
\value{
R: soil respiration rate, in umol m-2 s-1. This R port always returns 0 (soil respiration is not simulated) - kept for interface compatibility with the original SCOPE model.
}
\description{
soil respiration
}
\examples{
soil_respiration(20)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
