% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tav.R
\name{tav}
\alias{tav}
\title{Stern's formula in Lekner & Dorf (1988) gives reflectance for alfa = 90 degrees}
\usage{
tav(alfa, nr)
}
\arguments{
\item{alfa}{numeric. Maximum incidence solid angle, in degrees.}

\item{nr}{numeric. Leaf refractive index.}
}
\value{
angles: average transmissivity of a dielectric surface.
}
\description{
Stern's formula in Lekner & Dorf (1988) gives reflectance for alfa = 90 degrees
}
\examples{
tav(59, 1.4)
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Christiaan van der Tol (Original version in Matlab)
}\if{html}{\out{</div>}}

Carlos Camino (Ported version into R)
}
